update: modified files

+++ cms/serv/sql/migration/011_create_order_items.sql
@@ -11,9 +11,9 @@ CREATE TABLE order_items (
     sku_image         TEXT NOT NULL DEFAULT '',                       -- 图片
     sku_desc TEXT NOT NULL DEFAULT '',                                -- SKU 描述（如颜色、尺寸等）
     quantity SMALLINT NOT NULL DEFAULT 0,                             -- 数量
-    price NUMERIC(10, 2) NOT NULL DEFAULT 0.00,                       -- 单价（下单时快照）
-    total_amount NUMERIC(12, 2) NOT NULL DEFAULT 0.00,                -- 小计 = quantity * price
-    pay_amount NUMERIC(12, 2) NOT NULL DEFAULT 0.00,                  -- 小计 = quantity * price
+    price BIGINT NOT NULL DEFAULT 0,                    -- 单价（下单时快照）
+    total_amount BIGINT NOT NULL DEFAULT 0,             -- 小计 = quantity * price
+    pay_amount BIGINT NOT NULL DEFAULT 0,                 -- 小计 = quantity * price
     status SMALLINT NOT NULL DEFAULT 0,                               -- 状态（0: 待支付, 1: 已支付, 2: 已取消, 3: 已退款）
     cts             BIGINT DEFAULT (EXTRACT(EPOCH FROM NOW()) * 1000)::BIGINT,
     uts             BIGINT DEFAULT (EXTRACT(EPOCH FROM NOW()) * 1000)::BIGINT
